开始改bug了

+++ src/main/java/com/yunhuakeji/attendance/cache/ClockDaySettingCacheService.java
@@ -20,7 +20,7 @@ public class ClockDaySettingCacheService extends DataCacheService {
         List<ClockDaySetting> clockDaySettingList = clockDaySettingService.listAll();
         List<Integer> dayList = new ArrayList<>();
         for (ClockDaySetting setting : clockDaySettingList) {
-            Integer yearMonthDay = DateUtil.ymdToint(setting.getYearMonth(), setting.getDay());
+            int yearMonthDay = (int)DateUtil.ymdToint(setting.getYearMonth(), setting.getDay());
             dayList.add(yearMonthDay);
         }
         return dayList;
